favorites_fix_del

+++ backend/api/views.py
@@ -1,6 +1,6 @@
 from rest_framework import filters, status, viewsets, permissions
 from rest_framework.response import Response 
-from rest_framework.pagination import LimitOffsetPagination
+from rest_framework.pagination import LimitOffsetPagination, PageNumberPagination
 from djoser.views import UserViewSet as DjoserUserViewSet
 from django_filters.rest_framework import DjangoFilterBackend
 from django.shortcuts import get_object_or_404, redirect
@@ -103,7 +103,7 @@ class RecipeViewSet(viewsets.ModelViewSet):
     serializer_class = RecipeDetailSerializer
     filter_backends = (DjangoFilterBackend,)
     filterset_class = RecipeFilter
-    pagination_class = LimitOffsetPagination
+# pagination_class = PageNumberPagination
 
     def get_serializer_class(self):
         if self.request.method in ['POST', 'PUT', 'PATCH']:
@@ -113,25 +113,27 @@ class RecipeViewSet(viewsets.ModelViewSet):
     def get_permissions(self):
         if self.request.method in ['POST']:
             return [IsAuthenticated()]
+        elif self.action == 'favorite':  
+            return [IsAuthenticated()]  
         elif self.request.method in ['PUT', 'PATCH', 'DELETE']:
-            return [DeleteAndUdateOnlyAuthor()]  
-        return super().get_permissions()
+            return [DeleteAndUdateOnlyAuthor()]
+        return super().get_permissions() 
 
-    def list(self, request, *args, **kwargs):
-        queryset = self.filter_queryset(self.get_queryset())
+    # def list(self, request, *args, **kwargs):
+    #     queryset = self.filter_queryset(self.get_queryset())
 
-        page = self.paginate_queryset(queryset)
-        if page is not None:
-            serializer = self.get_serializer(page, many=True)
-            return self.get_paginated_response(serializer.data)
+    #     page = self.paginate_queryset(queryset)
+    #     if page is not None:
+    #         serializer = self.get_serializer(page, many=True)
+    #         return self.get_paginated_response(serializer.data)
 
-        serializer = self.get_serializer(queryset, many=True)
-        return Response({
-            "count": queryset.count(),
-            "next": self.get_next_link(),
-            "previous": self.get_previous_link(),
-            "results": serializer.data
-        })
+    #     serializer = self.get_serializer(queryset, many=True)
+    #     return Response({
+    #         "count": queryset.count(),
+    #         "next": self.get_next_link(),
+    #         "previous": self.get_previous_link(),
+    #         "results": serializer.data
+    #     })
 
     def create(self, request, *args, **kwargs):
         serializer = self.get_serializer(
@@ -186,28 +188,25 @@ class RecipeViewSet(viewsets.ModelViewSet):
             status=status.HTTP_200_OK,
         )
     
-    @action(detail=True, methods=['post'], permission_classes=[IsAuthenticated])
+    @action(detail=True, methods=['post', 'delete'], permission_classes=[IsAuthenticated])
     def favorite(self, request, pk=None):
         recipe = self.get_object()
-        favorite, created = Favorite.objects.get_or_create(user=request.user, recipe=recipe)
-
-        favorite_serializer = FavoriteSerializer(
-            favorite, context={'request': request}
-        )
-        if created:
-            return Response(favorite_serializer.data, status=status.HTTP_201_CREATED)
-        return Response({'status': 'already in favorites'}, status=status.HTTP_400_BAD_REQUEST)
-
-    @action(detail=True, methods=['delete'], permission_classes=[IsAuthenticated])
-    def unfavorite(self, request, pk=None):
-        recipe = self.get_object()
-        try:
-            favorite = Favorite.objects.get(user=request.user, recipe=recipe)
-            favorite.delete()
-            return Response({'status': 'removed from favorites'}, status=status.HTTP_204_NO_CONTENT)
-        except Favorite.DoesNotExist:
-            return Response({'status': 'not in favorites'}, status=status.HTTP_400_BAD_REQUEST)
-
+        
+        if request.method == 'POST':
+            # Добавление в избранное
+            favorite, created = Favorite.objects.get_or_create(user=request.user, recipe=recipe)
+            favorite_serializer = FavoriteSerializer(favorite, context={'request': request})
+            if created:
+                return Response(favorite_serializer.data, status=status.HTTP_201_CREATED)
+            return Response({'status': 'already in favorites'}, status=status.HTTP_400_BAD_REQUEST)
+        elif request.method == 'DELETE':
+            # Удаление из избранного
+            try:
+                favorite = Favorite.objects.get(user=request.user, recipe=recipe)
+                favorite.delete()
+                return Response({'status': 'removed from favorites'}, status=status.HTTP_204_NO_CONTENT)
+            except Favorite.DoesNotExist:
+                return Response({'status': 'not in favorites'}, status=status.HTTP_400_BAD_REQUEST)
 
     # def get_next_link(self):
     #     if self.paginator and self.paginator.get_next_link():
@@ -226,35 +225,35 @@ def recipe_short_link(request, hash):
     recipe = get_object_or_404(Recipe, short_link=hash)
     return redirect(f"/recipes/{recipe.id}/")
 
-class FavoritViewSet(viewsets.ViewSet):
-    permission_classes = [IsAuthenticated]
+# class FavoritViewSet(viewsets.ViewSet):
+#     permission_classes = [IsAuthenticated]
 
-    def create(self, request, recipe_id=None):
-        """Добавление рецепта в избранное."""
-        try:
-            recipe = Recipe.objects.get(id=recipe_id)
-        except Recipe.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+#     def create(self, request, recipe_id=None):
+#         """Добавление рецепта в избранное."""
+#         try:
+#             recipe = Recipe.objects.get(id=recipe_id)
+#         except Recipe.DoesNotExist:
+#             return Response(status=status.HTTP_404_NOT_FOUND)
 
-        favorite, created = Favorite.objects.get_or_create(user=request.user, recipe=recipe)
+#         favorite, created = Favorite.objects.get_or_create(user=request.user, recipe=recipe)
 
-        if created:
-            serializer = FavoriteSerializer(favorite)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        else:
-            return Response({"detail": "Рецепт уже добавлен в избранное."}, status=status.HTTP_400_BAD_REQUEST)
+#         if created:
+#             serializer = FavoriteSerializer(favorite)
+#             return Response(serializer.data, status=status.HTTP_201_CREATED)
+#         else:
+#             return Response({"detail": "Рецепт уже добавлен в избранное."}, status=status.HTTP_400_BAD_REQUEST)
 
-    def list(self, request):
-        """Просмотр списка избранных рецептов."""
-        favorites = Favorite.objects.filter(user=request.user)
-        serializer = FavoriteSerializer(favorites, many=True, context={'request': request})
-        return Response(serializer.data)
+#     def list(self, request):
+#         """Просмотр списка избранных рецептов."""
+#         favorites = Favorite.objects.filter(user=request.user)
+#         serializer = FavoriteSerializer(favorites, many=True, context={'request': request})
+#         return Response(serializer.data)
 
-    def destroy(self, request, recipe_id=None):
-        """Удаление рецепта из избранного."""
-        try:
-            favorite = Favorite.objects.get(user=request.user, recipe_id=recipe_id)
-            favorite.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        except Favorite.DoesNotExist:
-            return Response({"detail": "Рецепт не найден в избранном."}, status=status.HTTP_404_NOT_FOUND)
+#     def destroy(self, request, recipe_id=None):
+#         """Удаление рецепта из избранного."""
+#         try:
+#             favorite = Favorite.objects.get(user=request.user, recipe_id=recipe_id)
+#             favorite.delete()
+#             return Response(status=status.HTTP_204_NO_CONTENT)
+#         except Favorite.DoesNotExist:
+#             return Response({"detail": "Рецепт не найден в избранном."}, status=status.HTTP_404_NOT_FOUND)
